@@ -7,37 +7,49 @@
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
     "lint": "prettier --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-    "format": "prettier --write --plugin-search-dir=. ."
+    "format": "prettier --write --plugin-search-dir=. .",
+    "test:e2e": "playwright test",
+    "test:e2e:ui": "playwright test --ui",
+    "test:e2e:chromium": "playwright test --project=chromium",
+    "test:e2e:debug": "playwright test --debug",
+    "test:e2e:codegen": "playwright codegen",
+    "test:report": "playwright show-report"
   },
   "devDependencies": {
-    "@inqling/svelte-icons": "^3.1",
-    "@sveltejs/adapter-cloudflare": "^2.2.0",
-    "@sveltejs/kit": "^1.30.4",
-    "@tailwindcss/forms": "^0.5",
-    "@typescript-eslint/eslint-plugin": "^5.56",
-    "@typescript-eslint/parser": "^5.56",
-    "autoprefixer": "^10.4",
-    "eslint": "^8.36",
-    "eslint-config-prettier": "^8.8",
-    "eslint-plugin-svelte3": "^4.0",
+    "@playwright/test": "^1.51.1",
+    "@sveltejs/adapter-cloudflare": "^7.0.1",
+    "@sveltejs/kit": "^2.20.5",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/typography": "^0.5.16",
+    "@types/node": "^22.14.1",
+    "@typescript-eslint/eslint-plugin": "^8.29.1",
+    "@typescript-eslint/parser": "^8.29.1",
+    "autoprefixer": "^10.4.21",
+    "eslint": "^9.24.0",
+    "eslint-config-prettier": "^10.1.2",
+    "eslint-plugin-svelte": "^3.5.1",
     "just-debounce": "^1.1.0",
-    "mdsvex": "^0.10",
-    "postcss": "^8.4",
-    "postcss-load-config": "^4.0",
-    "prettier": "~2.8",
-    "prettier-plugin-svelte": "^2.9",
+    "lucide-svelte": "^0.487.0",
+    "mdsvex": "^0.12.3",
+    "prettier": "^3.5.3",
+    "prettier-plugin-svelte": "^3.3.3",
     "rehype-add-classes": "^1.0.0",
-    "rehype-autolink-headings": "^6.1.1",
-    "rehype-slug": "^5.1.0",
-    "rehype-urls": "^1.1.1",
-    "remark-abbr": "^1.4.1",
-    "svelte": "^3.57",
-    "svelte-check": "^3.1",
-    "svelte-preprocess": "^5.0",
-    "tailwindcss": "^3.2",
-    "tslib": "^2.5",
-    "typescript": "^4.9",
-    "vite": "^4.5"
+    "rehype-autolink-headings": "^7.1.0",
+    "rehype-slug": "^6.0.0",
+    "rehype-urls": "^1.2.0",
+    "remark-abbr": "^1.4.2",
+    "svelte": "^5.26.1",
+    "svelte-check": "^4.1.5",
+    "svelte-preprocess": "^6.0.3",
+    "tailwindcss": "^4.1.3",
+    "tslib": "^2.8.1",
+    "typescript": "^5.8.3",
+    "vite": "^6.2.6"
   },
-  "type": "module"
+  "type": "module",
+  "dependencies": {
+    "@sentry/sveltekit": "^9.13.0",
+    "@tailwindcss/vite": "^4.1.3"
+  }
 }
