@@ -1,7 +1,19 @@
-<script>
-  export let title = 'Article';
-  export let description = 'A worldwide open-source brewery dataset and API';
-  export let coverImageUrl = '/obdb-og.png';
+<script lang="ts">
+  import MarkdownContent from '$lib/components/MarkdownContent.svelte';
+
+  interface Props {
+    title?: string;
+    description?: string;
+    coverImageUrl?: string;
+    children?: import('svelte').Snippet;
+  }
+
+  let {
+    title = 'Article',
+    description = 'A worldwide open-source brewery dataset and API',
+    coverImageUrl = '/obdb-og.png',
+    children,
+  }: Props = $props();
 </script>
 
 <svelte:head>
@@ -11,16 +23,10 @@
   <meta property="og:title" content={title} />
   <meta property="og:description" content={description} />
   <meta property="og:image" content={coverImageUrl} />
-
-  <meta name="twitter:card" content="summary_large_image" />
-  <meta name="twitter:site" content="@openbrewerydb" />
-  <meta name="twitter:title" content={title} />
-  <meta name="twitter:description" content={description} />
-  <meta name="twitter:image" content={coverImageUrl} />
 </svelte:head>
 
-<article>
+<MarkdownContent>
   <h1>{title}</h1>
-  <slot />
-  <a href="/blog">Back to articles</a>
-</article>
+  {@render children?.()}
+  <a href="/blog" class="inline-block mt-6">Back to articles</a>
+</MarkdownContent>
