@@ -2,12 +2,17 @@
   import smallLogo from '$lib/assets/obdb-logo-sm.png';
   import NavMenuItem from './NavMenuItem.svelte';
 
-  export let toggleMenu: () => void;
+  interface Props {
+    toggleMenu: () => void;
+  }
+
+  let { toggleMenu }: Props = $props();
 </script>
 
 <nav
   class="relative flex items-center justify-between sm:h-10 md:justify-center"
   aria-label="Global"
+  data-testid="header-nav"
 >
   <div class="flex items-center flex-1 md:absolute md:inset-y-0 md:left-0">
     <div class="flex items-center justify-between w-full md:w-auto">
@@ -16,14 +21,14 @@
         <img
           class="h-12 w-auto sm:h-16"
           src={smallLogo}
-          alt="Open Brewery DB Logo"
+          alt="Open Brewery DB Logo Link to Home"
         />
       </a>
       <div class="-mr-2 flex items-center md:hidden">
         <button
-          on:click={toggleMenu}
+          onclick={toggleMenu}
           type="button"
-          class="bg-amber-200 rounded-md p-2 inline-flex items-center justify-center text-gray-600 hover:text-gray-800 hover:bg-amber-300 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-yellow-500"
+          class="bg-amber-200 rounded-md p-2 inline-flex items-center justify-center text-gray-700 hover:text-gray-900 hover:bg-amber-300 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-amber-500 shadow-sm hover:shadow"
           aria-expanded="false"
         >
           <span class="sr-only">Open main menu</span>
@@ -60,7 +65,7 @@
     <span class="inline-flex rounded-md shadow">
       <a
         href="http://eepurl.com/dBjS0j"
-        class="inline-flex items-center px-4 py-2 border border-transparent text-base font-medium rounded-md text-yellow-600 bg-white hover:bg-gray-50"
+        class="inline-flex items-center px-4 py-2 border border-transparent text-base font-medium rounded-md text-amber-600 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-amber-500 shadow-sm hover:shadow transition-shadow duration-200"
       >
         Newsletter Signup
       </a>
