@@ -2,27 +2,25 @@
   import Footer from '$lib/components/Footer.svelte';
   import MobileNav from '$lib/components/MobileNav.svelte';
   import Nav from '$lib/components/Nav.svelte';
-
-  import '../css/app.css';
-
-  let showMenu = false;
-
-  function toggleMenu() {
-    showMenu = !showMenu;
+  import '../styles/tailwind.css';
+  interface Props {
+    children?: import('svelte').Snippet;
   }
+
+  let { children }: Props = $props();
+
+  let showMenu: boolean = $state(false);
+
+  // Move function declaration to top level to fix ESLint error
+  const toggleMenu = () => {
+    showMenu = !showMenu;
+  };
 </script>
 
 <svelte:head>
   <meta property="og:type" content="article" />
   <meta
     property="og:image"
-    content="https://www.openbrewerydb.org/obdb-og.png"
-  />
-
-  <meta name="twitter:card" content="summary_large_image" />
-  <meta name="twitter:site" content="@openbrewerydb" />
-  <meta
-    name="twitter:image"
     content="https://www.openbrewerydb.org/obdb-og.png"
   />
 </svelte:head>
@@ -39,7 +37,7 @@
         </div>
       </div>
       <main class="mt-12 mx-auto max-w-7xl px-4">
-        <slot />
+        {@render children?.()}
       </main>
     </div>
   </div>
