@@ -5,29 +5,33 @@
   import Pagination from '$lib/components/Pagination.svelte';
   import { locationString } from '$lib/utils';
 
-  /** @type {import('./$types').PageData} */
-  export let data;
 
-  $: breweries = data.breweries ?? [];
-  $: country = data.country ?? '';
-  $: state = data.state ?? '';
-  $: meta = data.meta;
-  $: pageTitle = `Breweries in ${locationString({
+  /**
+   * @typedef {Object} Props
+   * @property {import('./$types').PageData} data
+   */
+
+  /** @type {Props} */
+  let { data } = $props();
+
+  let breweries = $derived(data.breweries ?? []);
+  let country = $derived(data.country ?? '');
+  let state = $derived(data.state ?? '');
+  let meta = $derived(data.meta);
+  let pageTitle = $derived(`Breweries in ${locationString({
     country,
     state,
-  })} | Open Brewery DB`;
-  $: pageDescription = `Breweries in ${locationString({
+  })} | Open Brewery DB`);
+  let pageDescription = $derived(`Breweries in ${locationString({
     country,
     state,
-  })} - Page ${meta.page}`;
+  })} - Page ${meta.page}`);
 </script>
 
 <svelte:head>
   <title>{pageTitle}</title>
   <meta property="og:title" content={pageTitle} />
   <meta property="og:description" content={pageDescription} />
-  <meta name="twitter:title" content={pageTitle} />
-  <meta name="twitter:description" content={pageDescription} />
 </svelte:head>
 
 <div class="">
