@@ -1,12 +1,20 @@
-<script>
+<script lang="ts">
   import NavMenuItem from './NavMenuItem.svelte';
 </script>
 
 <footer class="border-t-4 border-amber-100 pt-6 mt-6 mx-6">
   <div class="max-w-7xl mx-auto overflow-hidden">
-    <nav class="-mx-5 -my-2 flex flex-wrap justify-center" aria-label="Footer">
+    <nav
+      class="-mx-5 -my-2 flex flex-wrap justify-center"
+      aria-label="Footer"
+      data-testid="footer-nav"
+    >
       <div class="px-5 py-2">
         <NavMenuItem href="/">Home</NavMenuItem>
+      </div>
+
+      <div class="px-5 py-2">
+        <NavMenuItem href="/breweries">Breweries</NavMenuItem>
       </div>
 
       <div class="px-5 py-2">
@@ -27,18 +35,19 @@
     </nav>
     <div class="mt-8 flex justify-center space-x-6">
       <a
-        href="https://twitter.com/openbrewerydb"
+        href="https://bsky.app/profile/openbrewerydb.org"
         class="text-amber-700 hover:text-amber-800"
+        aria-label="GitHub"
       >
-        <span class="sr-only">Twitter</span>
+        <span class="sr-only">Bluesky</span>
         <svg
           class="h-6 w-6"
           fill="currentColor"
-          viewBox="0 0 24 24"
+          viewBox="0 0 600 540"
           aria-hidden="true"
         >
           <path
-            d="M8.29 20.251c7.547 0 11.675-6.253 11.675-11.675 0-.178 0-.355-.012-.53A8.348 8.348 0 0022 5.92a8.19 8.19 0 01-2.357.646 4.118 4.118 0 001.804-2.27 8.224 8.224 0 01-2.605.996 4.107 4.107 0 00-6.993 3.743 11.65 11.65 0 01-8.457-4.287 4.106 4.106 0 001.27 5.477A4.072 4.072 0 012.8 9.713v.052a4.105 4.105 0 003.292 4.022 4.095 4.095 0 01-1.853.07 4.108 4.108 0 003.834 2.85A8.233 8.233 0 012 18.407a11.616 11.616 0 006.29 1.84"
+            d="m135.72 44.03c66.496 49.921 138.02 151.14 164.28 205.46 26.262-54.316 97.782-155.54 164.28-205.46 47.98-36.021 125.72-63.892 125.72 24.795 0 17.712-10.155 148.79-16.111 170.07-20.703 73.984-96.144 92.854-163.25 81.433 117.3 19.964 147.14 86.092 82.697 152.22-122.39 125.59-175.91-31.511-189.63-71.766-2.514-7.3797-3.6904-10.832-3.7077-7.8964-0.0174-2.9357-1.1937 0.51669-3.7077 7.8964-13.714 40.255-67.233 197.36-189.63 71.766-64.444-66.128-34.605-132.26 82.697-152.22-67.108 11.421-142.55-7.4491-163.25-81.433-5.9562-21.282-16.111-152.36-16.111-170.07 0-88.687 77.742-60.816 125.72-24.795z"
           />
         </svg>
       </a>
@@ -46,6 +55,7 @@
       <a
         href="https://github.com/openbrewerydb"
         class="text-amber-700 hover:text-amber-800"
+        aria-label="GitHub"
       >
         <span class="sr-only">GitHub</span>
         <svg
@@ -65,6 +75,7 @@
       <a
         href="https://discord.gg/3G3syaD"
         class="text-amber-700 hover:text-amber-800"
+        aria-label="Discord"
       >
         <span class="sr-only">Discord</span>
         <svg
@@ -81,8 +92,8 @@
         </svg>
       </a>
     </div>
-    <p class="mt-8 text-center text-base">
-      &copy; 2023 Open Brewery DB. All rights reserved.
+    <p class="mt-8 text-center text-base text-gray-700">
+      &copy; 2025 Open Brewery DB. All rights reserved.
     </p>
   </div>
 </footer>
