@@ -1,7 +1,4 @@
-import * as env from '$env/static/public';
-
-export const API_URL: string =
-  env?.PUBLIC_API_URL || 'https://api.openbrewerydb.org/v1';
+export const API_URL = 'https://api.openbrewerydb.org/v1';
 
 export const mappings = {
   countries: {
