@@ -1,8 +1,17 @@
+import { sentrySvelteKit } from '@sentry/sveltekit';
+import { defineConfig } from 'vite';
 import { sveltekit } from '@sveltejs/kit/vite';
+import tailwindcss from '@tailwindcss/vite';
 
-/** @type {import('vite').UserConfig} */
-const config = {
-  plugins: [sveltekit()],
-};
-
-export default config;
+export default defineConfig({
+  plugins: [
+    sentrySvelteKit({
+      sourceMapsUploadOptions: {
+        org: 'open-brewery-db',
+        project: 'obdb-sveltekit',
+      },
+    }),
+    sveltekit(),
+    tailwindcss(),
+  ],
+});
